Starting implementation with Nicole
array_for_nicole.dart --> where nicole will update the method there to send the 2-d array to change, and also where to put the updated 2-d array (aka, next moves) to be eventually sent to the pico to change the LEDs and such. Effectively, the logic for implementing AI and chess logic for actual turns in chess.